+++ crates/sncast/src/helpers/mod.rs
@@ -1,2 +1,3 @@
+pub mod build;
 pub mod constants;
 pub mod scarb_utils;