+++ crates/sncast/tests/e2e/declare.rs
@@ -1,6 +1,7 @@
 use crate::helpers::constants::{CONTRACTS_DIR, URL};
 use crate::helpers::fixtures::{
-    duplicate_directory_with_salt, get_accounts_path, get_transaction_hash, get_transaction_receipt,
+    copy_directory_to_tempdir, duplicate_directory_with_salt, get_accounts_path,
+    get_transaction_hash, get_transaction_receipt,
 };
 use indoc::indoc;
 use snapbox::cmd::{cargo_bin, Command};
@@ -32,12 +33,158 @@ async fn test_happy_case() {
     let snapbox = Command::new(cargo_bin!("sncast"))
         .current_dir(contract_path.path())
         .args(args);
-    let output = snapbox.assert().success().get_output().stdout.clone();
+
+    let output = snapbox.assert().success().get_output().clone();
+    let output = output.stdout.clone();
 
     let hash = get_transaction_hash(&output);
     let receipt = get_transaction_receipt(hash).await;
 
     assert!(matches!(receipt, Declare(_)));
+
+    fs::remove_dir_all(contract_path).unwrap();
+}
+
+#[tokio::test]
+async fn test_worskpaces_package_specified() {
+    let contract_path =
+        copy_directory_to_tempdir(CONTRACTS_DIR.to_string() + "/hello_workspaces_cast");
+
+    let accounts_json_path = get_accounts_path("tests/data/accounts/accounts.json");
+    let args = vec![
+        "--url",
+        URL,
+        "--accounts-file",
+        accounts_json_path.as_str(),
+        "--account",
+        "user8",
+        "--int-format",
+        "--json",
+        "--package",
+        "addition_cast",
+        "declare",
+        "--contract-name",
+        "AdditionContract",
+        "--max-fee",
+        "99999999999999999",
+    ];
+
+    let snapbox = Command::new(cargo_bin!("sncast"))
+        .current_dir(contract_path.path())
+        .args(args);
+
+    let output = snapbox.assert().success().get_output().clone();
+    let output = output.stdout.clone();
+
+    let hash = get_transaction_hash(&output);
+    let receipt = get_transaction_receipt(hash).await;
+
+    assert!(matches!(receipt, Declare(_)));
+
+    fs::remove_dir_all(contract_path).unwrap();
+}
+
+#[tokio::test]
+async fn test_worskpaces_package_not_specified() {
+    let contract_path =
+        copy_directory_to_tempdir(CONTRACTS_DIR.to_string() + "/hello_workspaces_cast");
+
+    let accounts_json_path = get_accounts_path("tests/data/accounts/accounts.json");
+    let args = vec![
+        "--url",
+        URL,
+        "--accounts-file",
+        accounts_json_path.as_str(),
+        "--account",
+        "user8",
+        "--int-format",
+        "declare",
+        "--contract-name",
+        "AdditionContract",
+        "--max-fee",
+        "99999999999999999",
+    ];
+
+    let snapbox = Command::new(cargo_bin!("sncast"))
+        .current_dir(contract_path.path())
+        .args(args);
+
+    snapbox.assert().success().stderr_matches(indoc! {r"
+        command: declare
+        error: Failed to find artifacts in starknet_artifacts.json file[..]
+    "});
+
+    fs::remove_dir_all(contract_path).unwrap();
+}
+
+#[tokio::test]
+async fn test_worskpaces_package_specified_virtual() {
+    let contract_path =
+        copy_directory_to_tempdir(CONTRACTS_DIR.to_string() + "/hello_workspaces_virtual");
+
+    let accounts_json_path = get_accounts_path("tests/data/accounts/accounts.json");
+    let args = vec![
+        "--url",
+        URL,
+        "--accounts-file",
+        accounts_json_path.as_str(),
+        "--account",
+        "user8",
+        "--int-format",
+        "--json",
+        "--package",
+        "addition_virtual",
+        "declare",
+        "--contract-name",
+        "AdditionContract",
+        "--max-fee",
+        "99999999999999999",
+    ];
+
+    let snapbox = Command::new(cargo_bin!("sncast"))
+        .current_dir(contract_path.path())
+        .args(args);
+
+    let output = snapbox.assert().success().get_output().clone();
+    let output = output.stdout.clone();
+
+    let hash = get_transaction_hash(&output);
+    let receipt = get_transaction_receipt(hash).await;
+
+    assert!(matches!(receipt, Declare(_)));
+
+    fs::remove_dir_all(contract_path).unwrap();
+}
+
+#[tokio::test]
+async fn test_worskpaces_package_not_specified_virtual() {
+    let contract_path =
+        copy_directory_to_tempdir(CONTRACTS_DIR.to_string() + "/hello_workspaces_virtual");
+
+    let accounts_json_path = get_accounts_path("tests/data/accounts/accounts.json");
+    let args = vec![
+        "--url",
+        URL,
+        "--accounts-file",
+        accounts_json_path.as_str(),
+        "--account",
+        "user8",
+        "--int-format",
+        "declare",
+        "--contract-name",
+        "AdditionContract",
+        "--max-fee",
+        "99999999999999999",
+    ];
+
+    let snapbox = Command::new(cargo_bin!("sncast"))
+        .current_dir(contract_path.path())
+        .args(args);
+
+    snapbox.assert().failure().stderr_matches(indoc! {r"
+        Error: could not determine which package to work on
+        help: use the `--package` option to specify the package
+    "});
 
     fs::remove_dir_all(contract_path).unwrap();
 }
@@ -91,7 +238,6 @@ async fn wrong_contract_name_passed() {
 }
 
 #[test_case("/build_fails", "../../accounts/accounts.json" ; "when wrong cairo contract")]
-#[test_case("/", "../accounts/accounts.json" ; "when Scarb.toml does not exist")]
 fn scarb_build_fails(contract_path: &str, accounts_file_path: &str) {
     let args = vec![
         "--url",
@@ -110,8 +256,31 @@ fn scarb_build_fails(contract_path: &str, accounts_file_path: &str) {
         .args(args);
 
     snapbox.assert().stderr_matches(indoc! {r"
-        command: declare
-        error: Failed to build contracts with Scarb: `scarb` exited with error
+        Error: Failed to build contracts with Scarb
+        ...
+    "});
+}
+
+#[test_case("/", "../accounts/accounts.json" ; "when Scarb.toml does not exist")]
+fn scarb_build_fails_no_toml(contract_path: &str, accounts_file_path: &str) {
+    let args = vec![
+        "--url",
+        URL,
+        "--accounts-file",
+        accounts_file_path,
+        "--account",
+        "user1",
+        "declare",
+        "--contract-name",
+        "BuildFails",
+    ];
+
+    let snapbox = Command::new(cargo_bin!("sncast"))
+        .current_dir(CONTRACTS_DIR.to_string() + contract_path)
+        .args(args);
+
+    snapbox.assert().stderr_matches(indoc! {"
+        Error: No manifest found at the given path[..]
     "});
 }
 
@@ -166,9 +335,9 @@ fn scarb_no_sierra_artifact() {
         .current_dir(CONTRACTS_DIR.to_string() + "/no_sierra")
         .args(args);
 
-    snapbox.assert().success().stderr_matches(indoc! {r"
-        command: declare
-        [..]Make sure you have enabled sierra code generation in Scarb.toml[..]
+    snapbox.assert().failure().stderr_matches(indoc! {r"
+        [..]Make sure you have enabled sierra code generation in Scarb.toml
+        ...
     "});
 }