+++ crates/sncast/tests/e2e/main_tests.rs
@@ -1,11 +1,14 @@
 use crate::helpers::constants::{ACCOUNT, ACCOUNT_FILE_PATH, CONTRACTS_DIR, URL};
-use crate::helpers::fixtures::{duplicate_directory_with_salt, from_env, get_keystores_path};
+use crate::helpers::fixtures::{
+    create_basic_scarb_environment, duplicate_directory_with_salt, from_env, get_keystores_path,
+};
 use crate::helpers::runner::runner;
 use indoc::indoc;
 use snapbox::cmd::{cargo_bin, Command};
 use sncast::helpers::constants::KEYSTORE_PASSWORD_ENV_VAR;
 use std::env;
 use std::fs;
+use tempfile::TempDir;
 
 #[tokio::test]
 async fn test_happy_case_from_scarb() {
@@ -119,7 +122,11 @@ async fn test_happy_case_mixed() {
 
 #[tokio::test]
 async fn test_missing_account() {
+    let manifest_path = create_basic_scarb_environment();
+
     let args = vec![
+        "--path-to-scarb-toml",
+        &manifest_path,
         "--accounts-file",
         ACCOUNT_FILE_PATH,
         "--url",
@@ -138,7 +145,11 @@ async fn test_missing_account() {
 
 #[tokio::test]
 async fn test_missing_url() {
+    let manifest_path = create_basic_scarb_environment();
+
     let args = vec![
+        "--path-to-scarb-toml",
+        &manifest_path,
         "--accounts-file",
         ACCOUNT_FILE_PATH,
         "--account",
@@ -157,7 +168,11 @@ async fn test_missing_url() {
 
 #[tokio::test]
 async fn test_inexistent_keystore() {
+    let manifest_path = create_basic_scarb_environment();
+
     let args = vec![
+        "--path-to-scarb-toml",
+        &manifest_path,
         "--url",
         URL,
         "--keystore",
@@ -176,7 +191,20 @@ async fn test_inexistent_keystore() {
 
 #[tokio::test]
 async fn test_keystore_account_required() {
+    let temp_dir = TempDir::new().expect("Unable to create a temporary directory");
+    let manifest_path = temp_dir.path().join("./Scarb.toml");
+    let manifest_path = manifest_path.to_str().unwrap();
+
+    fs_extra::file::copy(
+        "tests/data/files/noconfig_Scarb.toml",
+        manifest_path,
+        &fs_extra::file::CopyOptions::new().overwrite(true),
+    )
+    .unwrap();
+
     let args = vec![
+        "--path-to-scarb-toml",
+        manifest_path,
         "--url",
         URL,
         "--keystore",
@@ -195,7 +223,20 @@ async fn test_keystore_account_required() {
 
 #[tokio::test]
 async fn test_keystore_inexistent_account() {
+    let temp_dir = TempDir::new().expect("Unable to create a temporary directory");
+    let manifest_path = temp_dir.path().join("./Scarb.toml");
+    let manifest_path = manifest_path.to_str().unwrap();
+
+    fs_extra::file::copy(
+        "tests/data/files/noconfig_Scarb.toml",
+        manifest_path,
+        &fs_extra::file::CopyOptions::new().overwrite(true),
+    )
+    .unwrap();
+
     let args = vec![
+        "--path-to-scarb-toml",
+        manifest_path,
         "--url",
         URL,
         "--keystore",